Add tracking of source line execution counts
Changes JaCoCo's underlying data structure from a `boolean[]` to an `int[]`
such that we increment values in this array whenever instructions are being
executed. For incrementing the integer value, we cast to a double and back to
prevent integers from overflowing, essentially capping execution counts to 
`Integer.MAX_VALUE`.

The execution counts for multiple instructions on a single source code line are
combined by taking their maximum value. This line execution count is then
exposed in the HTML reports on hover and in XML reports using the new `ec`
attribute.